Some changes

+++ tasks/baldin_a_word_count/common/include/common.hpp
@@ -8,8 +8,8 @@
 namespace baldin_a_word_count {
 
 using InType = std::string;
-using OutType = size_t;
-using TestType = std::tuple<std::string, size_t>;
+using OutType = long long;
+using TestType = std::tuple<std::string, long long>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace baldin_a_word_count
